fixed: raised Exception on a server does not map to exception on a client
Current version does not "understand" the following server output (SOAP 1.1):
...
    <env:Fault xmlns:n1="http://schemas.xmlsoap.org/soap/encoding/"
        env:encodingStyle="http://schemas.xmlsoap.org/soap/encoding/">
      <faultcode>env:Server</faultcode>
      <faultstring xsi:type="xsd:string">UssWsApiException</faultstring>
      <faultactor xsi:type="xsd:string">AuthInterfaceServlet</faultactor>
      <detail xmlns:n2="urn:uss-wsapi:Auth">
        <n2:UssWsApiException xsi:type="n2:UssWsApiException">
          <errorCode xsi:type="xsd:int">50000</errorCode>
          <errorDescription xsi:type="xsd:string">FAILURE</errorDescription>
          <message xsi:type="xsd:string">Authentication failure</message>
        </n2:UssWsApiException>
      </detail>
    </env:Fault>
...

It raises SOAP::Fault exception instead of mapped in registry
UssWsApiException. This is because "detail" does not have appropriate
"type" (as "n2:UssWsApiException" does), so it cannot be mapped to a
correct exception. But "detail" HAS "type" by default (see
SOAPUnknown.decode_tag_by_type,
lib/soap/encodingstyle/soapHandler.rb:457)

Let's make more checks. If "detail"'s type is not "detail" itself, if it
has children and first child's type is a SOAP::SOAPStruct then try to
map those child to a type (exception). Then if that type is an
exception raise it but not "fault".

Also, increase version to 1.5.10 to be able to detect THIS version of a gem.